Move pin-button script, style and template files into components/pin-button  🐿 v2.9.0
