+++ components/pin-button/index.js
