+++ components/pin-button/main.scss
